AutoSkill Activity/Fragments were named weird
